Make the default "default-compiler" be the nixpkgs default
The NixOS project keep a handy binary cache of pkgs.haskellPackages,
which is currently an alias for pkgs.haskell.packages.ghc802.

If styx selects ghc801 by default, and the current "unstable" channel
of nixpkgs is used, then there will be a lot of rebuilding required to
prepare a cabal sandbox.

This change defers the default compiler selection to be whatever is
used by the selected nixpkgs.